+++ common/src/main/java/traben/entity_model_features/models/animation/EMFAnimation.java
@@ -4,22 +4,24 @@ import it.unimi.dsi.fastutil.floats.FloatConsumer;
 import it.unimi.dsi.fastutil.objects.Object2FloatOpenHashMap;
 import it.unimi.dsi.fastutil.objects.Object2ObjectLinkedOpenHashMap;
 import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
-import traben.entity_model_features.EMF;
 import traben.entity_model_features.models.EMFModelPart;
 import traben.entity_model_features.models.animation.math.MathComponent;
 import traben.entity_model_features.models.animation.math.MathExpressionParser;
 import traben.entity_model_features.models.animation.math.variables.EMFModelOrRenderVariable;
 import traben.entity_model_features.models.animation.math.variables.factories.GlobalVariableFactory;
-import traben.entity_model_features.utils.EMFUtils;
 
 import java.util.UUID;
+
+import static traben.entity_model_features.models.animation.math.MathValue.FALSE;
 
 public class EMFAnimation {
 
     public final String animKey;
     public final String expressionString;
     public final String modelName;
-    public final boolean isVariable;
+    public boolean isVariable(){
+        return handleVariableResult != null;
+    }
     private final EMFModelPart partToApplyTo;
     private final EMFModelOrRenderVariable modelOrRenderVariableToChange;
     private final Object2FloatOpenHashMap<UUID> prevResult = new Object2FloatOpenHashMap<>();
@@ -27,7 +29,7 @@ public class EMFAnimation {
     public Object2ObjectLinkedOpenHashMap<String, EMFAnimation> emfAnimationVariables = null;
     public Object2ObjectOpenHashMap<String, EMFModelPart> allPartsBySingleAndFullHeirachicalId = null;
     private MathComponent emfCalculator = MathExpressionParser.NULL_EXPRESSION;
-    private FloatConsumer handleVariableResult = null;
+    private final FloatConsumer handleVariableResult;
 
     public EMFAnimation(EMFModelPart partToApplyTo,
                         EMFModelOrRenderVariable modelOrRenderVariableToChange,
@@ -37,34 +39,57 @@ public class EMFAnimation {
     ) {
         this.modelName = modelName;
         this.animKey = animKey;
-        isVariable = animKey.startsWith("var") || animKey.startsWith("global_var");
-        if (isVariable) {
-            handleVariableResult = animKey.startsWith("global_var") ?
-                    value -> GlobalVariableFactory.setGlobalVariable(animKey, value)
-            :       value -> EMFAnimationEntityContext.setEntityVariable(animKey, value);
+        boolean animKeyIsBoolean = (animKey.startsWith("global_varb") || animKey.startsWith("varb"));
+
+        if (animKey.startsWith("global_var")){
+            //global
+            if (animKeyIsBoolean){
+                //boolean
+                handleVariableResult = value -> GlobalVariableFactory.setGlobalVariable(animKey, Float.isInfinite(value) ? value : FALSE);
+            }else{
+                //float
+                handleVariableResult = value -> GlobalVariableFactory.setGlobalVariable(animKey, Float.isInfinite(value) ? 0 : value);
+            }
+        }else if (animKey.startsWith("var")){
+            //entity
+            if (animKeyIsBoolean){
+                //boolean
+                handleVariableResult = value -> EMFAnimationEntityContext.setEntityVariable(animKey, Float.isInfinite(value) ? value : FALSE);
+            }else{
+                //float
+                handleVariableResult = value -> EMFAnimationEntityContext.setEntityVariable(animKey, Float.isInfinite(value) ? 0 : value);
+            }
+        } else {
+            handleVariableResult = null;
         }
-        this.modelOrRenderVariableToChange = isVariable ? null : modelOrRenderVariableToChange;
+
+        this.modelOrRenderVariableToChange = isVariable() ? null : modelOrRenderVariableToChange;
         this.partToApplyTo = partToApplyTo;
 
         float defaultValue;
-        if (this.modelOrRenderVariableToChange != null) {
-            if (partToApplyTo == null) {
-                if (this.modelOrRenderVariableToChange.isRenderVariable()) {
-                    defaultValue = this.modelOrRenderVariableToChange.getValue();
-                } else {
-                    if (EMF.config().getConfig().logModelCreationData)
-                        EMFUtils.log("null part for " + animKey);
-                    defaultValue = 0;
-                }
-            } else {
-                defaultValue = this.modelOrRenderVariableToChange.getValue(partToApplyTo);
-            }
-        } else {
-            defaultValue = 0;
-        }
-        prevResult.defaultReturnValue(defaultValue);
+//        if (this.modelOrRenderVariableToChange != null) {
+//            if (partToApplyTo == null) {
+//                if (this.modelOrRenderVariableToChange.isRenderVariable()) {
+//                    defaultValue = this.modelOrRenderVariableToChange.getValue();
+//                } else {
+//                    if (EMF.config().getConfig().logModelCreationData)
+//                        EMFUtils.log("null part for " + animKey);
+//                    defaultValue = 0;
+//                }
+//            } else {
+//                defaultValue = this.modelOrRenderVariableToChange.getValue(partToApplyTo);
+//            }
+//        } else {
+            defaultValue = animKeyIsBoolean ||
+                    (modelOrRenderVariableToChange != null && modelOrRenderVariableToChange.isBoolean())
+                    ? FALSE : 0;
+//        }
+        this.defaultValue = defaultValue;
+        prevResult.defaultReturnValue(this.defaultValue);
         expressionString = initialExpression;
     }
+
+    private final float defaultValue;
 
 
     @Override
@@ -85,7 +110,7 @@ public class EMFAnimation {
 
     public float getLastResultOnly() {
         if (EMFAnimationEntityContext.getEMFEntity() == null) {
-            return 0;
+            return defaultValue;
         }
         return prevResult.getFloat(EMFAnimationEntityContext.getEMFEntity().etf$getUuid());
 
@@ -94,7 +119,7 @@ public class EMFAnimation {
     public float getResultViaCalculate() {
         UUID id = EMFAnimationEntityContext.getEMFEntity() == null ? null : EMFAnimationEntityContext.getEMFEntity().etf$getUuid();
         if (id == null) {
-            return 0;
+            return defaultValue;
         }
 
         float result = calculatorRun();
@@ -106,8 +131,8 @@ public class EMFAnimation {
 
     private float calculatorRun() {
         float result = emfCalculator.getResult();
-        if (Float.isNaN(result) || Float.isInfinite(result) ||  Math.abs(result) == Float.MIN_VALUE) {
-            return 0;
+        if (Float.isNaN(result) || Math.abs(result) == Float.MIN_VALUE) {
+            return defaultValue;
         } else {
             return result;
         }
@@ -115,27 +140,25 @@ public class EMFAnimation {
 
 
     public void calculateAndSet() {
-        if (EMFAnimationEntityContext.isLODSkippingThisFrame()) {
-            handleResult(getLastResultOnly());
-        } else {
-            calculateAndSetPostLod();
-        }
+            if (EMFAnimationEntityContext.isLODSkippingThisFrame()) {
+                handleResultNonVariable(getLastResultOnly());
+            } else {
+                calculateAndSetPostLod();
+            }
     }
 
     private void calculateAndSetPostLod() {
-        if (isVariable) {
-            if (handleVariableResult != null) {
-                handleVariableResult.accept(getResultViaCalculate());
-            } else {
-                EMFUtils.logError(animKey + ": variable did not have result handler in: " + modelName);
-            }
+        if (isVariable()) {
+            handleVariableResult.accept(getResultViaCalculate());
         } else {
-            handleResult(getResultViaCalculate());
+            handleResultNonVariable(getResultViaCalculate());
         }
     }
 
-    private void handleResult(float result) {
+    private void handleResultNonVariable(float result) {
         if (modelOrRenderVariableToChange != null) {
+            //todo potentially could validate boolean and non boolean values here, but that could represent a performance hit
+            //modelOrRenderVariableToChange.isBoolean()... etc
             modelOrRenderVariableToChange.setValue(partToApplyTo, result);
         }
     }