+++ common/src/main/java/traben/entity_model_features/models/animation/math/variables/VariableRegistry.java
@@ -131,6 +131,7 @@ public final class VariableRegistry {
         registerSimpleBoolVariable("is_sprinting", EMFAnimationEntityContext::isSprinting);
         registerSimpleBoolVariable("is_tamed", EMFAnimationEntityContext::isTamed);
         registerSimpleBoolVariable("is_wet", EMFAnimationEntityContext::isWet);
+        registerSimpleBoolVariable("is_jumping", EMFAnimationEntityContext::isJumping);
 
         //context variables
         // these variables require a context to be created, and are not constants