+++ common/src/main/java/traben/entity_model_features/utils/EMFEntity.java
@@ -58,4 +58,5 @@ public interface EMFEntity extends ETFEntity {
     Object2FloatOpenHashMap<String> emf$getVariableMap();
 
 
+
 }